+++ ryven/ironflow/__init__.py
@@ -1,4 +1,4 @@
-__all__ = ['gui', 'flow_canvas', 'canvas_widgets', 'layouts', 'models', 'node_interface']
+__all__ = ['gui', 'flow_canvas', 'canvas_widgets', 'layouts', 'model.py', 'node_interface']
 
 from ryven.ironflow.gui import GUI
 from ryven.NENV import Node, NodeInputBP, NodeOutputBP, dtypes
